+++ src/main/java/com/raoulvdberge/refinedstorage/item/CapabilityProviderEnergy.java
@@ -1,9 +1,6 @@
 package com.raoulvdberge.refinedstorage.item;
 
 import com.raoulvdberge.refinedstorage.integration.forgeenergy.ItemEnergyForge;
-import com.raoulvdberge.refinedstorage.integration.tesla.IntegrationTesla;
-import com.raoulvdberge.refinedstorage.integration.tesla.ItemEnergyTesla;
-import net.darkhax.tesla.capability.TeslaCapabilities;
 import net.minecraft.item.ItemStack;
 import net.minecraft.util.EnumFacing;
 import net.minecraftforge.common.capabilities.Capability;
@@ -22,26 +19,11 @@ public class CapabilityProviderEnergy implements ICapabilityProvider {
 
     @Override
     public boolean hasCapability(@Nonnull Capability<?> capability, @Nullable EnumFacing facing) {
-        return capability == CapabilityEnergy.ENERGY ||
-            (IntegrationTesla.isLoaded() && (capability == TeslaCapabilities.CAPABILITY_HOLDER || capability == TeslaCapabilities.CAPABILITY_CONSUMER));
+        return capability == CapabilityEnergy.ENERGY;
     }
 
     @Override
     public <T> T getCapability(@Nonnull Capability<T> capability, @Nullable EnumFacing facing) {
-        if (capability == CapabilityEnergy.ENERGY) {
-            return CapabilityEnergy.ENERGY.cast(new ItemEnergyForge(stack, ItemEnergyItem.CAPACITY));
-        }
-
-        if (IntegrationTesla.isLoaded()) {
-            if (capability == TeslaCapabilities.CAPABILITY_HOLDER) {
-                return TeslaCapabilities.CAPABILITY_HOLDER.cast(new ItemEnergyTesla(stack));
-            }
-
-            if (capability == TeslaCapabilities.CAPABILITY_CONSUMER) {
-                return TeslaCapabilities.CAPABILITY_CONSUMER.cast(new ItemEnergyTesla(stack));
-            }
-        }
-
-        return null;
+        return capability == CapabilityEnergy.ENERGY ? CapabilityEnergy.ENERGY.cast(new ItemEnergyForge(stack, ItemEnergyItem.CAPACITY)) : null;
     }
 }